requirements.txt
python-telegram-bot>=20.0
selenium>=4.0.0
webdriver-manager>=3.8.0
python-dotenv>=0.19.0
